Port repair_observation_vote_cache script into RepairObservationVoteCacheJob
Part of the crontab-to-Solid-Queue audit (issue #4726, Bucket A). Brings the
Ruby logic directly into the job class rather than having the job shell out
to the script, and schedules it via config/recurring.yml (daily at 5:28am,
matching the old cron cadence) on the :maintenance queue.

Also fixes a pre-existing bug surfaced by the new test: send_alert took
rows.first(tally[:repaired]) - an arbitrary prefix of the full stale-rows
list ordered by created_at, not the rows actually classified :repaired. When
some stale-looking rows turn out :unchanged after recompute, that prefix
slice could name the wrong observation(s) in the alert email. Now tracks the
actually-repaired rows directly and alerts on those.

Deletes script/repair_observation_vote_cache.rb - no other references to it
remain in the repo.
